refactor all tasklet users into other APIs
During the work to modernize the tasklet API, there was a request to just entirely
remove it. This is much more work than just making the API safer,
so it is considered the "next step" in dealing with tasklets.

In the meantime, comments were added to suggest people quit using tasklets in new code. ;)

https://lore.kernel.org/driverdev-devel/20200716081538.2sivhkj4hcyrusem@linutronix.de/

These are a place holder for the series that will be sent upstream.

Signed-off-by: Allen Pais <apais@linux.microsoft.com>